requirements.txt
playwright
requests
beautifulsoup4
lxml
pandas
openpyxl
python-dotenv
fake-useragent
google-genai
mcp[cli]
